+++ R/FroeseWs.R
@@ -8,7 +8,6 @@
 #' 
 #' If the \code{col.pop} argument is set equal to one of these palettes -- \dQuote{rich}, \dQuote{cm}, \dQuote{default}, \dQuote{grey}, \dQuote{gray}, \dQuote{heat}, \dQuote{jet}, \dQuote{rainbow}, \dQuote{topo}, or \dQuote{terrain} -- and the \code{order.pop=TRUE} then the populations plotted should form a general color gradient from smallest to largest weight in the initial length category.  This will make it easier to identify populations that \dQuote{cross over} other populations.
 #' 
-#' @aliases FroeseWs plot.FroeseWs coef.FroeseWs
 #' @param log.a A numeric vector that contains the \eqn{log_{10}(a)} values for the population of length-weight regression equations.
 #' @param b A numeric vector that contains the b values for the population of length-weight regression equations
 #' @param x An object saved from the \code{FrowseWs()} call (i.e., of class \code{Froese}).
@@ -31,7 +30,7 @@
 #'
 #' @references Froese, R.  2006.  Cube law, condition factor and weight-length relationships: history, meta-analysis and recommendations.  Journal of Applied Ichthyology 22:241-253.
 #' 
-#' @export
+#' @aliases FroeseWs plot.FroeseWs coef.FroeseWs
 #' 
 #' @keywords manip hplot
 #' 
@@ -39,7 +38,7 @@
 #' #See examples in RuffeWs.
 #' 
 #' @rdname FroeseWs
-#' @export FroeseWs
+#' @export
 FroeseWs <- function(log.a,b) {
   mn.log.a <- mean(log.a)
   gm.a <- 10^mn.log.a
@@ -50,8 +49,7 @@ FroeseWs <- function(log.a,b) {
 }
 
 #' @rdname FroeseWs
-#' @method coef FroeseWs
-#' @S3method coef FroeseWs
+#' @export
 coef.FroeseWs <- function(object,...) {
   res <- c(object$gm.a,object$mn.b)
   names(res) <- c("gm.a","mn.b")
@@ -59,29 +57,40 @@ coef.FroeseWs <- function(object,...) {
 }
 
 #' @rdname FroeseWs
-#' @method plot FroeseWs
-#' @S3method plot FroeseWs
-plot.FroeseWs <- function(x,min,max,what=c("both","raw","log"),col.pop="rich",lwd.pop=1,lty.pop=1,order.pop=TRUE,col.Ws="black",lwd.Ws=3,lty.Ws=1,...) {
+#' @export
+plot.FroeseWs <- function(x,min,max,what=c("both","raw","log"),
+                          col.pop="rich",lwd.pop=1,lty.pop=1,order.pop=TRUE,
+                          col.Ws="black",lwd.Ws=3,lty.Ws=1,...) {
   object <- x
   what <- match.arg(what)
   if (col.pop %in% paletteChoices()) col.pop <- chooseColors(col.pop,length(object$b))
   len <- seq(min,max,length.out=200)
-  for (i in 1:length(object$log.a)) {                                     # predict weight at each 1-cm midpoint for each population
+  # predict weight at each 1-cm midpoint for each population
+  for (i in 1:length(object$log.a)) {                                     
     w <- 10^(object$log.a[i]+object$b[i]*log10(len))
-    if (i==1) pred.w <- w                                          # store in a matrix called pred.w
+    # store in a matrix called pred.w
+    if (i==1) pred.w <- w                                          
     else pred.w <- cbind(pred.w,w)
   }
-  colnames(pred.w) <- seq(1:dim(pred.w)[2])                        # rename columns to numbers that correspond to populations
+  # rename columns to numbers that correspond to populations
+  colnames(pred.w) <- seq(1:dim(pred.w)[2])                        
   if (order.pop) pred.w <- pred.w[,order(pred.w[1,])]
-  if (what=="both") old.par <- par(mar=c(3.5,3.5,1,1), mfcol=c(1,2), mgp=c(2,0.75,0))
-    else old.par <- par(mar=c(3.5,3.5,1,1), mgp=c(2,0.75,0))
-  on.exit(par(old.par))
+  if (what=="both") old.par <- graphics::par(mar=c(3.5,3.5,1,1),
+                                             mfcol=c(1,2),mgp=c(2,0.75,0))
+    else old.par <- graphics::par(mar=c(3.5,3.5,1,1), mgp=c(2,0.75,0))
+  on.exit(graphics::par(old.par))
   if (what=="raw" | what=="both") {
-    matplot(len,pred.w,type="l",lty=lty.pop,lwd=lwd.pop,col=col.pop,xlab="Length (mm)",ylab="Weight (g)")
-    curve(object$gm.a*x^object$mn.b,min(len),max(len),lty=lty.Ws,lwd=lwd.Ws,col=col.Ws,add=TRUE)
+    graphics::matplot(len,pred.w,type="l",lty=lty.pop,lwd=lwd.pop,col=col.pop,
+                      xlab="Length (mm)",ylab="Weight (g)")
+    graphics::curve(object$gm.a*x^object$mn.b,min(len),max(len),
+                    lty=lty.Ws,lwd=lwd.Ws,col=col.Ws,add=TRUE)
   }
   if (what=="log" | what=="both") {
-    matplot(log10(len),log10(pred.w),type="l",lty=lty.pop,lwd=lwd.pop,col=col.pop,xlab="log10(Length (mm))",ylab="log10(Weight (g))")
-    curve(log10(object$gm.a)+object$mn.b*x,min(log10(len)),max(log10(len)),lty=lty.Ws,lwd=lwd.Ws,col=col.Ws,add=TRUE)
+    graphics::matplot(log10(len),log10(pred.w),type="l",
+                      lty=lty.pop,lwd=lwd.pop,col=col.pop,
+                      xlab="log10(Length (mm))",ylab="log10(Weight (g))")
+    graphics::curve(log10(object$gm.a)+object$mn.b*x,
+                    min(log10(len)),max(log10(len)),
+                    lty=lty.Ws,lwd=lwd.Ws,col=col.Ws,add=TRUE)
   }
 }