+++ R/WalleyeGerowLW.R
@@ -12,12 +12,10 @@
 #'    \item{len}{Total length (mm) of fish.}
 #'    \item{wt}{Weight (g) of fish.} 
 #'  }
-#' 
-#' @concept Condition 'Relative Weight' 'Standard Weight' 'Length-Weight'
 #'
 #' @seealso \code{\link{rlp}}, \code{\link{emp}}, and \code{\link{wsValidate}}.
 #'
-#' @source Directly from Ken Gerow.  Summarized in Gerow, K.G., R.C. Anderson-Sprecher, and W.A. Hubert.  2005.  A new method to compute standard-weight equations that reduces length-related bias.  North American Journal of Fisheries Management 25:1288-1300.
+#' @source Directly from Ken Gerow. Summarized in Gerow, K.G., R.C. Anderson-Sprecher, and W.A. Hubert. 2005. A new method to compute standard-weight equations that reduces length-related bias. North American Journal of Fisheries Management 25:1288-1300.
 #'
 #' @keywords datasets
 #'